requirements.txt
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
ultralytics==8.0.196
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
PyYAML==6.0.1 